projects/BinaryGrace-FlowPay/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
web3==6.11.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
pydantic==2.5.0
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
